# libs/shared/api-client-core/src/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: string;
};

export type NotePageDto = {
    items: Array<{
        id: string;
        title: string;
        body: string;
        version: number;
        createdAt: string;
        updatedAt: string;
    }>;
    nextCursor: string | null;
};

export type NoteDto = {
    id: string;
    title: string;
    body: string;
    version: number;
    createdAt: string;
    updatedAt: string;
};

export type CreateNoteDto = {
    title: string;
    body?: string;
};

export type UpdateNoteDto = {
    title?: string;
    body?: string;
    version: number;
};

export type DemoItemPageDto = {
    items: Array<{
        id: string;
        title: string;
        createdAt: string;
        updatedAt: string;
    }>;
    nextCursor: string | null;
};

export type CreateDemoItemDto = {
    title: string;
};

export type DemoItemDto = {
    id: string;
    title: string;
    createdAt: string;
    updatedAt: string;
};

export type AppControllerGetDataData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api';
};

export type AppControllerGetDataResponses = {
    200: unknown;
};

export type WhoamiControllerWhoamiData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/whoami';
};

export type WhoamiControllerWhoamiResponses = {
    200: unknown;
};

export type NotesControllerListData = {
    body?: never;
    path?: never;
    query?: {
        cursor?: string;
        limit?: number;
    };
    url: '/api/v1/notes';
};

export type NotesControllerListResponses = {
    200: NotePageDto;
};

export type NotesControllerListResponse = NotesControllerListResponses[keyof NotesControllerListResponses];

export type NotesControllerCreateData = {
    body: CreateNoteDto;
    path?: never;
    query?: never;
    url: '/api/v1/notes';
};

export type NotesControllerCreateResponses = {
    201: NoteDto;
};

export type NotesControllerCreateResponse = NotesControllerCreateResponses[keyof NotesControllerCreateResponses];

export type NotesControllerRemoveData = {
    body?: never;
    path: {
        id: string;
    };
    query?: never;
    url: '/api/v1/notes/{id}';
};

export type NotesControllerRemoveResponses = {
    204: void;
};

export type NotesControllerRemoveResponse = NotesControllerRemoveResponses[keyof NotesControllerRemoveResponses];

export type NotesControllerFindData = {
    body?: never;
    path: {
        id: string;
    };
    query?: never;
    url: '/api/v1/notes/{id}';
};

export type NotesControllerFindResponses = {
    200: NoteDto;
};

export type NotesControllerFindResponse = NotesControllerFindResponses[keyof NotesControllerFindResponses];

export type NotesControllerUpdateData = {
    body: UpdateNoteDto;
    path: {
        id: string;
    };
    query?: never;
    url: '/api/v1/notes/{id}';
};

export type NotesControllerUpdateResponses = {
    200: NoteDto;
};

export type NotesControllerUpdateResponse = NotesControllerUpdateResponses[keyof NotesControllerUpdateResponses];

export type DemoItemsControllerListData = {
    body?: never;
    path?: never;
    query?: {
        limit?: number;
        cursor?: string;
    };
    url: '/api/v1/demo-items';
};

export type DemoItemsControllerListResponses = {
    200: DemoItemPageDto;
};

export type DemoItemsControllerListResponse = DemoItemsControllerListResponses[keyof DemoItemsControllerListResponses];

export type DemoItemsControllerCreateData = {
    body: CreateDemoItemDto;
    path?: never;
    query?: never;
    url: '/api/v1/demo-items';
};

export type DemoItemsControllerCreateResponses = {
    201: DemoItemDto;
};

export type DemoItemsControllerCreateResponse = DemoItemsControllerCreateResponses[keyof DemoItemsControllerCreateResponses];